Actually, this is not the right way to do this. We'll merge piecewise from trunk instead.

+++ src/mlpack/tests/sgd_test.cpp
@@ -49,7 +49,7 @@ BOOST_AUTO_TEST_CASE(GeneralizedRosenbrockTest)
 
     BOOST_REQUIRE_SMALL(result, 1e-10);
     for (size_t j = 0; j < i; ++j)
-      BOOST_REQUIRE_CLOSE(coordinates[j], (double) 1.0, 1e-3);
+      BOOST_REQUIRE_CLOSE(coordinates[j], 1.0, 1e-3);
   }
 }
